@ash2k 2023-12-11
I've added a new benchmark that exercises the `NewSetWithSortableFiltered()` function directly. Here are the results:

```
main:
BenchmarkNewSetWithSortableFiltered-10    	 8248227	       138.4 ns/op	     448 B/op	       1 allocs/op

branch:
BenchmarkNewSetWithSortableFiltered-10    	 6625119	       177.0 ns/op	     448 B/op	       1 allocs/op
```
Because we are testing sorting, actual performance depends on the data being sorted. But this branch performs 1 allocation too.