@MrAlias 2024-01-03
```console
$ go1.20 version
go version go1.20 linux/amd64

$ go1.20 test -run='^$' -bench=NewSetWithSortableFiltered
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/attribute
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkNewSetWithSortableFiltered-8   	 2496835	       480.1 ns/op	     584 B/op	       4 allocs/op
PASS
```

On main:

```console
go1.20 test -run='^$' -bench=NewSetWithSortableFiltered
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/attribute
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkNewSetWithSortableFiltered-8   	 4108585	       283.9 ns/op	     448 B/op	       1 allocs/op
PASS
```