@pellared 2024-01-25
On branch:

```
$ go version
go version go1.21.4 linux/amd64

$ go test -run=^$ -bench=NewSetWithSortableFiltered
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/attribute
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz
BenchmarkNewSetWithSortableFiltered-16           5002983               239.6 ns/op           448 B/op          1 allocs/op
PASS
ok      go.opentelemetry.io/otel/attribute      1.448s
```

I think we can consider this PR when we drop support for Go 1.20.