@MrAlias 2024-02-28
This should be ready to merge now that https://github.com/open-telemetry/opentelemetry-go/pull/4967 is merged.

Please update to remove the separate code paths for versions of Go less than 1.21 as they are no longer supported.