@ash2k 2024-03-06
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/attribute
```

New code:

```
BenchmarkFiltering
BenchmarkFiltering/NoFilter
BenchmarkFiltering/NoFilter/Set.Filter
BenchmarkFiltering/NoFilter/Set.Filter-10         	571478235	         2.061 ns/op	       0 B/op	       0 allocs/op
BenchmarkFiltering/NoFilter/NewSetWithFiltered
BenchmarkFiltering/NoFilter/NewSetWithFiltered-10 	 1000000	      1087 ns/op	    3584 B/op	       2 allocs/op
BenchmarkFiltering/NoFiltered
BenchmarkFiltering/NoFiltered/Set.Filter
BenchmarkFiltering/NoFiltered/Set.Filter-10       	 3128260	       382.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkFiltering/NoFiltered/NewSetWithFiltered
BenchmarkFiltering/NoFiltered/NewSetWithFiltered-10         	  967354	      1264 ns/op	    3584 B/op	       2 allocs/op
BenchmarkFiltering/Filtered
BenchmarkFiltering/Filtered/Set.Filter
BenchmarkFiltering/Filtered/Set.Filter-10                   	 1326331	       870.4 ns/op	    1856 B/op	       2 allocs/op
BenchmarkFiltering/Filtered/NewSetWithFiltered
BenchmarkFiltering/Filtered/NewSetWithFiltered-10           	 1217002	       985.0 ns/op	      64 B/op	       1 allocs/op
BenchmarkFiltering/AllDropped
BenchmarkFiltering/AllDropped/Set.Filter
BenchmarkFiltering/AllDropped/Set.Filter-10                 	 1418766	       844.3 ns/op	    1792 B/op	       1 allocs/op
BenchmarkFiltering/AllDropped/NewSetWithFiltered
BenchmarkFiltering/AllDropped/NewSetWithFiltered-10         	 2279227	       525.0 ns/op	       0 B/op	       0 allocs/op
```

Old code:

```
BenchmarkFiltering
BenchmarkFiltering/NoFilter
BenchmarkFiltering/NoFilter/Set.Filter
BenchmarkFiltering/NoFilter/Set.Filter-10         	556945123	         2.053 ns/op	       0 B/op	       0 allocs/op
BenchmarkFiltering/NoFilter/NewSetWithFiltered
BenchmarkFiltering/NoFilter/NewSetWithFiltered-10 	 1274888	       943.1 ns/op	    3585 B/op	       2 allocs/op
BenchmarkFiltering/NoFiltered
BenchmarkFiltering/NoFiltered/Set.Filter
BenchmarkFiltering/NoFiltered/Set.Filter-10       	 3141591	       381.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkFiltering/NoFiltered/NewSetWithFiltered
BenchmarkFiltering/NoFiltered/NewSetWithFiltered-10         	 1000000	      1106 ns/op	    3585 B/op	       2 allocs/op
BenchmarkFiltering/Filtered
BenchmarkFiltering/Filtered/Set.Filter
BenchmarkFiltering/Filtered/Set.Filter-10                   	 1377650	       865.5 ns/op	    1856 B/op	       2 allocs/op
BenchmarkFiltering/Filtered/NewSetWithFiltered
BenchmarkFiltering/Filtered/NewSetWithFiltered-10           	 1629787	       733.1 ns/op	      64 B/op	       1 allocs/op
BenchmarkFiltering/AllDropped
BenchmarkFiltering/AllDropped/Set.Filter
BenchmarkFiltering/AllDropped/Set.Filter-10                 	 1416115	       844.3 ns/op	    1792 B/op	       1 allocs/op
BenchmarkFiltering/AllDropped/NewSetWithFiltered
BenchmarkFiltering/AllDropped/NewSetWithFiltered-10         	 3032091	       388.7 ns/op	       0 B/op	       0 allocs/op
```